fix(search): add title and id to ordering
When using the currentness ordering, the relevant criterium to sort by
is the date-based rank. But until now in some cases inconsistent results
coud be retrieved through the /initial and the /search endpoints, as the
implicit database ordering was handled differently in some cases for
entries with the same rank. With this fix, the ordering should always be
consistend, as not only the title, but also the unique id is part of the
ordering.

Refs: #2222