fix: extend media file field length
While files in Portfolio are also limited to 255 characters, Showroom
does store the whole URI to the file on Portfolio. So in some cases
Portfolio cannot push a medium, if the filename already is close to the
max length. With this increase to 350 characters max length we have
sufficient reserve to preprend the actual filename with the repository's
base URI for the files.