fix: use static description for field help_text
The problem before this fix was, that if you changed the config in a way
that the list of available plugin changes, also the help_text will
change, which will result in Django telling us that there is a change
not reflected in the existing migrations. After a discussion in the
backend team we decided to make this static again and refer to the docs.