fix: add extra check for relations integrity error
This extra check was not there, because it should not be needed. Using
the .add() method of the RelatedManager takes care of potential
duplicates by doing a SELECT before the potential INSERT. I could not
even intentionally reproduce this that way. Only by subsituting the .add
with a Relation.objects.bulk_create() and adding a list with the same
two relations in it.

Nevertheless we encountered an IntegrityError here in production on
2024-01-05 (and probably once, or twice before). This might be an issue
with the underlying database.

The ideal solution would be to find the actual source of these rare
IntegrityErrors and fix it wherever it happens. Until then we now have
this extra check, so that Portfolio does not get a 500 when syncing
relations, and we get some more debug info.